+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change Log
 
 - Updated the search bar on all search screens to a new, modern design.
 - Updated the icon for pinning tabs to be a pushpin for better clarity.
+- Cleaned up the Settings screen by centralizing all session import actions on the Login screen.
 
 ## Version 0.9.6(90600)